[IMP] account{,_sequence}: merge
This module was added in stable but was never intended as a separate
module.
