--- addons/account_sequence/models/__init__.py
@@ -1,2 +1,0 @@
-from . import account_move
-from . import sequence_mixin
