--- addons/l10n_latam_account_sequence/__init__.py
@@ -1,1 +1,0 @@
-from . import models
