--- addons/l10n_latam_account_sequence/models/__init__.py
@@ -1,1 +1,0 @@
-from . import account_move
